Add response status to http method in SimulationController class

+++ src/main/java/com/simulation/simulationapp/controller/SimulationController.java
@@ -3,6 +3,8 @@ package com.simulation.simulationapp.controller;
 import com.simulation.simulationapp.model.Simulation;
 import com.simulation.simulationapp.service.SimulationService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -31,17 +33,25 @@ public class SimulationController {
     }
 
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public void addSimulation(@RequestBody Simulation simulation){
         simulationService.addSimulation(simulation);
     }
 
     @PutMapping("/{id}")
+    @ResponseStatus(HttpStatus.ACCEPTED)
     public void updateSimulation(@PathVariable("id") long idSimulation,@RequestBody Simulation updateSimulation){
         simulationService.updateSimulation(idSimulation,updateSimulation);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteSimulation(@PathVariable("id") long idSimulation){
-        simulationService.deleteSimulation(idSimulation);
+    public ResponseEntity<?> deleteSimulation(@PathVariable("id") long idSimulation){
+        boolean anyElementRemoved = simulationService.findSimulationById(idSimulation).isPresent();
+
+        if (anyElementRemoved) {
+            simulationService.deleteSimulation(idSimulation);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
